requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosmtplib==3.0.2
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
autopep8==2.3.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.2.1
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
openai==1.9.0
pillow==11.0.0
PyAudio==0.2.14
pycodestyle==2.12.1
pydantic==2.5.3
pydantic_core==2.14.6
pypiwin32==223
pyTelegramBotAPI==4.23.0
pyttsx3==2.90
pywin32==306
requests==2.31.0
schedule==1.2.2
sniffio==1.3.0
SpeechRecognition==3.10.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
